chore: create `sc.io`, move `sc.{read*,write*,external.exporting.*}` into it
