--- docs/external/exporting.md
@@ -1,14 +1,0 @@
-# Exporting
-
-```{eval-rst}
-.. module:: scanpy.external.exporting
-.. currentmodule:: scanpy.external
-```
-
-```{eval-rst}
-.. autosummary::
-   :toctree: generated/
-
-   exporting.spring_project
-   exporting.cellbrowser
-```
