+++ docs/release-notes/1.3.6.md
@@ -9,7 +9,7 @@
 #### Interactive exploration of analysis results through *manifold viewers*
 
 - CZI’s [cellxgene] directly reads `.h5ad` files {smaller}`the cellxgene developers`
-- the [UCSC Single Cell Browser] requires exporting via {func}`~scanpy.external.exporting.cellbrowser` {smaller}`M Haeussler`
+- the [UCSC Single Cell Browser] requires exporting via {func}`~scanpy.io.write_cellbrowser` (introduced here as `scanpy.external.exporting.write_cellbrowser`) {smaller}`M Haeussler`
 
 #### Code design
