+++ src/scanpy/__init__.py
@@ -9,24 +9,13 @@ from typing import TYPE_CHECKING
 from ._settings import Preset, Verbosity, settings  # isort: skip
 
 from anndata import AnnData, concat
-from anndata.io import (
-    read_csv,
-    read_excel,
-    read_h5ad,
-    read_hdf,
-    read_loom,
-    read_mtx,
-    read_text,
-    read_umi_tools,
-)
 
-from . import datasets, experimental, external, get, logging, metrics, queries
+from . import datasets, experimental, external, get, io, logging, metrics, queries
 from . import plotting as pl
 from . import preprocessing as pp
 from . import tools as tl
 from ._utils import annotate_doc_types
 from .neighbors import Neighbors
-from .readwrite import read, read_10x_h5, read_10x_mtx, read_visium, write
 
 if TYPE_CHECKING:
     from typing import Any
@@ -41,27 +30,15 @@ __all__ = [
     "experimental",
     "external",
     "get",
+    "io",
     "logging",
     "metrics",
     "pl",
     "pp",
     "queries",
-    "read",
-    "read_10x_h5",
-    "read_10x_mtx",
-    "read_csv",
-    "read_excel",
-    "read_h5ad",
-    "read_hdf",
-    "read_loom",
-    "read_mtx",
-    "read_text",
-    "read_umi_tools",
-    "read_visium",
     "set_figure_params",
     "settings",
     "tl",
-    "write",
 ]
 
 
@@ -74,13 +51,26 @@ sys.modules.update({f"{__name__}.{m}": globals()[m] for m in ["tl", "pp", "pl"]}
 
 
 def __getattr__(name: str) -> Any:
+    import anndata
+
+    from ._compat import warn
+
     if name == "__version__":
         from importlib.metadata import version
-
-        from ._compat import warn
 
         msg = "`__version__` is deprecated, use `importlib.metadata.version('scanpy')` instead"
         warn(msg, FutureWarning)
         return version("scanpy")
+
+    if name in {"read", "read_10x_h5", "read_10x_mtx", "write"} or (
+        name in io.__all__ and name in anndata.io.__all__
+    ):
+        msg = "Import from `scanpy.io` instead"
+        warn(msg, FutureWarning)
+        return getattr(io, name)
+    if name == "read_visium":
+        from .io._read import read_visium
+
+        return read_visium  # warns on call already
 
     raise AttributeError
