+++ src/scanpy/external/__init__.py
@@ -5,9 +5,10 @@ from __future__ import annotations
 import sys
 
 from .. import _utils
-from . import exporting, pl, pp, tl
+from . import _exporting as exporting  # noqa: F401
+from . import pl, pp, tl
 
 _utils.annotate_doc_types(sys.modules[__name__], "scanpy")
 del sys, _utils
 
-__all__ = ["exporting", "pl", "pp", "tl"]
+__all__ = ["pl", "pp", "tl"]
